refactor: drop pkg folder
