--- internal/doc.go
@@ -1,3 +1,0 @@
-// Package internal includes code that's not meant to be used by external
-// applications.
-package internal
