--- pkg/doc.go
@@ -1,2 +1,0 @@
-// Package pkg includes library code that's ok to use by external applications.
-package pkg
